# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API сервиса подачи заявок на 3D-печать МИЭМ НИУ ВШЭ.\\nАутентификация — OTP через email + JWT (Bearer token).",
        "title": "MIEM 3D Lab Print API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/api/admin/stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-stats"
                ],
                "summary": "Статистика (admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Начало периода (YYYY-MM-DD)",
                        "name": "date_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Конец периода (YYYY-MM-DD)",
                        "name": "date_to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_dto.StatsResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный период",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/admin/users": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-users"
                ],
                "summary": "Поиск пользователей по email (admin)",
                "parameters": [
                    {
                        "minLength": 3,
                        "type": "string",
                        "description": "Минимум 3 символа",
                        "name": "email",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_dto.AdminUsersResponse"
                        }
                    },
                    "400": {
                        "description": "Запрос слишком короткий",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/admin/users/{id}/role": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-users"
                ],
                "summary": "Изменить роль пользователя (admin)",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "ID пользователя",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новая роль",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_dto.SetRoleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_dto.SetRoleResponse"
                        }
                    },
                    "400": {
                        "description": "Недопустимое значение роли",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Нельзя снять роль с последнего администратора",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "miem_3d_lab_print_backend_src_internal_apierr.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "INTERNAL_ERROR"
                },
                "details": {
                    "type": "object"
                },
                "message": {
                    "type": "string",
                    "example": "Внутренняя ошибка сервера"
                }
            }
        },
        "miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.APIError"
                }
            }
        },
...
        "miem_3d_lab_print_backend_src_internal_dto.AdminUserItem": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "miem_3d_lab_print_backend_src_internal_dto.AdminUsersResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_dto.AdminUserItem"
                    }
                }
            }
        },
...
        "miem_3d_lab_print_backend_src_internal_dto.ByMaterialStat": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "material_name": {
                    "type": "string"
                }
            }
        },
...
        "miem_3d_lab_print_backend_src_internal_dto.SetRoleRequest": {
            "type": "object",
            "properties": {
                "role": {
                    "type": "string"
                }
            }
        },
        "miem_3d_lab_print_backend_src_internal_dto.SetRoleResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "miem_3d_lab_print_backend_src_internal_dto.StatsResponse": {
            "type": "object",
            "properties": {
                "avg_completion_hours": {
                    "type": "number"
                },
                "by_material": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_dto.ByMaterialStat"
                    }
                },
                "by_status_current": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "completed_count": {
                    "type": "integer"
                },
                "period": {
                    "type": "object",
                    "properties": {
                        "date_from": {
                            "type": "string"
                        },
                        "date_to": {
                            "type": "string"
                        }
                    }
                },
                "total_applications": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите \"Bearer \u003caccess_token\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}